rename EventSpeakers to EventNumberOfSpeakers
