+++ packages/react-components/src/index.ts
@@ -56,8 +56,8 @@ export {
   DropdownButton,
   ErrorCard,
   EventInfo,
+  EventNumberOfSpeakers,
   EventOwner,
-  EventSpeakers,
   EventTeams,
   EventTime,
   ExternalLink,
