+++ packages/react-components/src/molecules/index.ts
@@ -12,8 +12,8 @@ export { default as EventInfo } from './EventInfo';
 export { default as EventMaterialComingSoon } from './EventMaterialComingSoon';
 export { default as EventMaterialsUnavailable } from './EventMaterialsUnavailable';
 export { default as EventMaterialUnavailable } from './EventMaterialUnavailable';
+export { default as EventNumberOfSpeakers } from './EventNumberOfSpeakers';
 export { default as EventOwner } from './EventOwner';
-export { default as EventSpeakers } from './EventSpeakers';
 export { default as EventTeams } from './EventTeams';
 export { default as EventTime } from './EventTime';
 export { default as ExternalLink } from './ExternalLink';
